ToDo_Project/todo/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.5
django-filter==24.3
django-rest-swagger==2.2.0
djangorestframework==3.15.2
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
Markdown==3.7
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==24.2
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
setuptools==75.8.0
simplejson==3.19.3
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.3.0
